requirements.txt
Flask
tensorflow==2.19.0
flask-cors
numpy
Pillow